requirements.txt
torch==2.1.0
torchvision==0.16.0
numpy==1.26.0
requests==2.31.0
python-chess==1.999
zstandard==0.22.0
python-dotenv==1.0.0
tensorboard==2.15.1
tensorflow==2.15.0
pygame==2.5.2